tweaks to make fetching work via celery
